requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Deprecated==1.2.13
Django==3.2.9
django-elasticsearch-dsl==6.5.0
django-redis==5.2.0
djangorestframework==3.13.1
elasticsearch==6.8.0
elasticsearch-dsl==6.4.0
packaging==21.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
